refactor: 상대경로 제거 및 글자색 조정
